+++ src/test/java/kitchenpos/menu/domain/MenuGroupRepositoryTest.java
@@ -1,9 +1,7 @@
-package kitchenpos.menu.domain.dao;
+package kitchenpos.menu.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
